# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ticketmaster API
  version: "1.0.0"
  description: |
    Event discovery and ticketing API.

    ## Authentication

    `POST /api/login` (or `/api/admin/login`) returns a token. Send it as
    `Authorization: Bearer <token>` on any endpoint marked with a padlock.
    Tokens expire 24 hours after they are issued.

    Endpoints fall into three tiers:

    * **public** — discovery reads, registration, login, password recovery
    * **user** — any signed-in account; a user only ever sees their own bookings
    * **admin** — an account with `role: admin`

    Every write to a discovery resource is admin-only, including the create
    routes that share a path with a public read.

    ## Partial updates

    `PUT` and `PATCH` behave identically: the body is decoded over the stored
    record, so send only the fields that change. Server-owned fields (`id`,
    and `ticketsSold` on an event) are ignored if present in a body.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same origin as this page
...
security: []
paths:
...
  /api/register:
    post:
      tags: [Auth]
      summary: Register a user account
      description: |
        `role` in the body is ignored — an account created here is always a
        normal user. Admin accounts come from `/api/admin/register`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, example: Ann Smith }
                email: { type: string, format: email, example: ann@example.com }
                password: { type: string, format: password, example: pw123456 }
      responses:
        "201":
          description: Account created. The password hash is never returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: name, email and password are required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Email already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/login:
    post:
      tags: [Auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/logout:
    post:
      tags: [Auth]
      summary: Sign out
      description: |
        Revokes the token the request was made with, so it stops working
        immediately rather than lingering until it expires. Only the calling
        session ends — the same account signed in on another device stays
        signed in.

        Not rate-limited: ending your own session must always be possible.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Session revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password-reset token
      description: |
        Always answers 200 with the same message whether or not the address has
        an account, so the endpoint cannot be used to discover who is
        registered. Admin accounts are excluded from the flow.

        No mail service is configured, so the token is written to the server
        log. Outside production (`ENV` set to development/dev/local/test) it is
        also returned in the response as `resetToken` to make local testing
        workable — note that this does reveal whether the account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Request accepted (identical whether or not the account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  resetToken:
                    type: string
                    description: Non-production only
                  note: { type: string }
        "400":
          description: email is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password using a reset token
      description: |
        The token is single-use and expires one hour after it was issued.
        Requesting a new one invalidates any earlier token. A successful reset
        also revokes every existing login session for that account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, description: The token from /api/forgot-password }
                newPassword: { type: string, format: password, minLength: 6 }
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "password updated, please sign in again" }
        "400":
          description: Missing fields, password too short, or an invalid/used/expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ------------------------------------------------------------ admin: auth
  /api/admin/register:
    post:
      tags: [Admin — accounts]
      summary: Create an admin account
      description: |
        Requires the shared secret from `ADMIN_REGISTRATION_KEY`, sent either as
        the `X-Admin-Key` header or as `adminKey` in the body. Registration is
        disabled (503) while that variable is unset.
      parameters:
        - name: X-Admin-Key
          in: header
          schema: { type: string }
          description: The registration secret. Alternatively send `adminKey` in the body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, format: password }
                adminKey: { type: string, description: Used when the header is absent }
      responses:
        "201":
          description: Admin account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: Invalid admin key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Email already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: ADMIN_REGISTRATION_KEY is not set, so registration is disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/admin/login:
    post:
      tags: [Admin — accounts]
      summary: Sign in as an admin
      description: |
        A non-admin account gets the same 401 as a wrong password, so this
        endpoint does not reveal which emails belong to admins.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: Invalid credentials, or the account is not an admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/admin/me:
    get:
      tags: [Admin — accounts]
      summary: The signed-in admin
      description: Useful for validating a stored token on page load.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The admin account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ------------------------------------------------------- admin: analytics
  /api/admin/analytics:
    get:
      tags: [Admin — bookings]
      summary: Dashboard figures
      description: |
        Both dashboard series in one response: revenue per event and tickets
        sold per category. Aggregated by the database, so the client does not
        need to read the booking collection to chart it.

        Confirmed bookings only — a cancelled booking has had its seats
        released, so counting it would overstate revenue and demand.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: topEvents
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
          description: How many events to return, highest revenue first. Capped at 50.
      responses:
        "200":
          description: Dashboard series
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Analytics" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # -------------------------------------------------------- admin: check-in
  /api/admin/tickets/check-in:
    post:
      tags: [Admin — bookings]
      summary: Admit a scanned ticket
      description: |
        Consumes the `ticketCode` a ticket's QR encodes and admits its holder.

        The first scan succeeds and stamps `checkedInAt`; every later scan of
        the same ticket is refused as `already_used`, so one ticket admits one
        time. The stamp is applied with a single conditional update, so two
        doors scanning simultaneously cannot both admit.

        The code travels in the body rather than the path: it is the ticket's
        only credential, and URLs end up in access logs and browser history.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: The ticketCode from the booking, as carried by the QR.
                  example: 9f3c1d7a5b2e480c9a6f1e2d3c4b5a60
      responses:
        "200":
          description: Admitted — first scan of a valid ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckIn" }
              example:
                result: valid
                message: admitted
        "400":
          description: No code supplied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: No ticket matches that code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckIn" }
              example:
                result: not_found
                message: no ticket matches that code
        "409":
          description: Already scanned, or the booking was cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckIn" }
              example:
                result: already_used
                message: this ticket has already been scanned

  # ----------------------------------------------------------- admin: users
  /api/admin/users:
    get:
      tags: [Admin — accounts]
      summary: List accounts
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: keyword, in: query, schema: { type: string }, description: Substring of name or email }
        - { name: role, in: query, schema: { type: string, enum: [user, admin] } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged accounts, without password hashes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedEnvelope"
                  - type: object
                    properties:
                      _embedded:
                        type: object
                        properties:
                          users:
                            type: array
                            items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/users/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Admin — accounts]
      summary: Get one account
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Admin — accounts]
      summary: Update an account (partial)
      description: |
        A `password` in the body is plaintext and is hashed before storage;
        omit it to leave the current password alone. An admin cannot remove
        their own admin role.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Missing email, or an attempt to self-demote
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Email already registered to another account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      tags: [Admin — accounts]
      summary: Update an account (identical to PUT)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Email already registered to another account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [Admin — accounts]
      summary: Delete an account
      description: |
        Refused with 409 while the user holds confirmed bookings. An admin
        cannot delete their own account (400).
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: "Deleted, along with that account's login tokens" }
        "400":
          description: Attempt to delete your own account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/InUse" }

  # -------------------------------------------------------- admin: bookings
  /api/admin/bookings:
    get:
      tags: [Admin — bookings]
      summary: List bookings across all users
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: query, schema: { type: string } }
        - { name: eventId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [confirmed, cancelled] } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged bookings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedEnvelope"
                  - type: object
                    properties:
                      _embedded:
                        type: object
                        properties:
                          bookings:
                            type: array
                            items: { $ref: "#/components/schemas/Booking" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/bookings/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Admin — bookings]
      summary: Get any booking
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Admin — bookings]
      summary: Delete a booking outright
      description: |
        Erases the record. A still-confirmed booking releases its tickets back
        to the event first, so an event is never left overcounted. Prefer
        `/cancel` when the history matters.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/bookings/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Admin — bookings]
      summary: Cancel any user's booking
      description: |
        Sets the status to `cancelled` and returns the tickets to the event,
        keeping the record for audit. Cancelling an already-cancelled booking
        is a safe no-op.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The cancelled booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------- user bookings
  /api/bookings:
    get:
      tags: [Bookings]
      summary: List your own bookings
      description: Returns a plain array, not the paged envelope used elsewhere.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Every booking belonging to the signed-in user
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Bookings]
      summary: Book tickets
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, quantity]
              properties:
                eventId: { type: string }
                quantity: { type: integer, minimum: 1, example: 2 }
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Event not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Event is not on sale, or not enough tickets remain
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/bookings/{id}/pay:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Bookings]
      summary: Confirm payment for a held booking
      description: |
        Turns a `pending` booking into a `confirmed` one, once the payment
        provider confirms the charge actually succeeded.

        The server asks the provider directly — the client only says which
        booking it is paying for. Trusting a client that claims "paid" would
        hand out tickets for free.

        Only meaningful when payments are enabled; answers 404 otherwise.
        Calling it twice is safe: an already-paid booking returns 200.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Paid, or already paid
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking: { $ref: "#/components/schemas/Booking" }
                  status: { type: string, enum: [paid, "already paid"] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402":
          description: The charge has not completed at the provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: No such booking, or payments are not enabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: The hold expired, or the booking is not awaiting payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: The payment provider could not be reached; the hold stands, so retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A token from `/api/login` or `/api/admin/login`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    TooManyRequests:
      description: |
        Rate limit exceeded for this endpoint and caller. The `Retry-After`
        header gives the number of seconds to wait.
      headers:
        Retry-After:
          description: Seconds until the next attempt is allowed
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "too many attempts, try again in 42 seconds" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      example: { error: event not found }
...
    User:
      type: object
      properties:
        id: { type: string, readOnly: true }
        name: { type: string, example: Ann Smith }
        email: { type: string, format: email }
        role:
          type: string
          enum: [user, admin]
          description: Set by the server from the endpoint used; ignored in request bodies on /api/register.
...